% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.r
\name{add.grid}
\alias{add.grid}
\title{Add grid lines for the plot}
\usage{
add.grid(x, adds.sq.legend = TRUE, draws.1.2ha = TRUE)
}
\arguments{
\item{x}{an \code{ogawa.plot} object.}

\item{adds.sq.legend}{if TRUE, adds an legend for sub-quadrats.}

\item{draws.1.2ha}{if TRUE, draws the 1.2ha core plot region.}
}
\description{
A low level graphic function draws grid lines for the ogawa plot map.
}
