% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{all_touching}
\alias{all_touching}
\title{Check all quadrats are touching each other}
\usage{
all_touching(..., corner = TRUE)
}
\arguments{
\item{...}{quadrat codes.}

\item{corner}{if TRUE, quadrats contiguous at their corners are also treated as
contiguous. Defaults to TRUE.}
}
\value{
Returns TRUE if all specified quadrats are contiguous.
}
\description{
Check all quadrats are touching each other
}
