% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{construct.q}
\alias{construct.q}
\title{Construct quadrat codes from elements.}
\usage{
construct.q(Q1, Q2, SQ1, SQ2)
}
\arguments{
\item{Q1}{vector of capital alphabet part of quadrat code.}

\item{Q2}{vector of first numerical part of quadrat code.}

\item{SQ1}{vector of lower alphabet part of quadrat code.}

\item{SQ2}{vector of second numerical part of quadrat code.}
}
\value{
a character vector of quadrat code(s).
}
\description{
Construct quadrat codes from elements (Q1, Q2, SQ1, SQ2).
}
