% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.r
\name{create.quadrat.map}
\alias{create.quadrat.map}
\title{Create a map of ogawa forest plot with colored quadrats}
\usage{
create.quadrat.map(Q, adds.sq.legend = TRUE, draws.1.2ha = TRUE, ...)
}
\arguments{
\item{Q}{quadrat codes to be filled.}

\item{adds.sq.legend}{if TRUE, adds an legend for sub-quadrats.}

\item{draws.1.2ha}{if TRUE, draws the 1.2ha core plot region.}

\item{...}{graphic parameters passed to \code{\link[graphics]{rect}} to draw quadrats.}
}
\description{
Create a map of ogawa forest plot with colored quadrats
}
