% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.r
\name{create_ogawa_plot}
\alias{create_ogawa_plot}
\title{Create new ogawa forest plot}
\usage{
create_ogawa_plot(
  xmin = 0,
  xmax = 300,
  ymin = 0,
  ymax = 200,
  label_pos_x = c(0.02, 0.05),
  label_pos_y = c(0.02, 0.06),
  ...
)
}
\arguments{
\item{xmin}{a minimum value of x coordinate.}

\item{xmax}{a maximum value of x coordinate.}

\item{ymin}{a minimum value of y coordinate.}

\item{ymax}{a maximum value of y coordinate.}

\item{label_pos_x}{a numeric vector with two elements. The first element denotes distance
(% of shorter axis) between small X axis labels and the grid and the
second denotes distance between large X axis labels and the grid.}

\item{label_pos_y}{a numeric vector with two elements. The first element denotes distance
(% of shorter axis) between small Y axis labels and the grid and the
second denotes distance between large Y axis labels and the grid.}
}
\value{
an \code{ogawa_plot} object which can be passed to other functions.
}
\description{
Create new plot for ogawa forest plot map.
}
\examples{
x <- create_ogawa_plot()
add_grid(x)

}
