% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.r
\name{create_point_map}
\alias{create_point_map}
\title{Create a map of ogawa forest plot with points}
\usage{
create_point_map(
  x,
  y,
  adds_sq_legend = TRUE,
  draws_1_2ha = TRUE,
  contour = FALSE,
  col = rgb(1, 0, 0, 0.5),
  xlim = c(0, 300),
  ylim = c(0, 200),
  ...
)
}
\arguments{
\item{x}{x coordinates.}

\item{y}{y coordinates.}

\item{adds_sq_legend}{if TRUE, adds an legend for sub-quadrats.}

\item{draws_1_2ha}{if TRUE, draws the 1.2ha core plot region.}

\item{contour}{if TRUE, draw contour.}

\item{col}{point color(s).}

\item{xlim}{range of x.}

\item{ylim}{range of y.}

\item{...}{graphic parameters passed to \code{\link[graphics]{rect}} to draw
quadrats.}
}
\description{
Create a map of ogawa forest plot with points
}
