% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.r
\name{create_quadrat_map}
\alias{create_quadrat_map}
\title{Create a map of ogawa forest plot with colored quadrats}
\usage{
create_quadrat_map(
  Q,
  adds_sq_legend = TRUE,
  draws_1_2ha = TRUE,
  contour = FALSE,
  col = rgb(1, 0, 0, 0.5),
  type = c("rect", "point"),
  jitter = FALSE,
  xlim = c(0, 300),
  ylim = c(0, 200),
  ...
)
}
\arguments{
\item{Q}{quadrat codes to be filled.}

\item{adds_sq_legend}{if TRUE, adds an legend for sub-quadrats.}

\item{draws_1_2ha}{if TRUE, draws the 1.2ha core plot region.}

\item{contour}{if TRUE, draw contour.}

\item{col}{fill color(s).}

\item{type}{= c("rect", "point")
if "rect", draws filled rectangular for each quadrat.
If "point", draws points for each quadrat.}

\item{jitter}{if TRUE, add jitter for coordinates of overlapping points.
Ignored when type is "rect".}

\item{xlim}{range of x.}

\item{ylim}{range of y.}

\item{...}{graphic parameters passed to \code{\link[graphics]{rect}} to draw
quadrats.}
}
\description{
Create a map of ogawa forest plot with colored quadrats
}
