% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.r
\name{fill.q}
\alias{fill.q}
\title{Fill quadrats}
\usage{
fill.q(Q, ...)
}
\arguments{
\item{Q}{quadrat codes.}

\item{...}{graphic parameters passed to \code{\link[graphics]{rect}}.}
}
\description{
Low-level graphic function fills quadrats specifyied by quadrat codes.
}
