% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{find_adjacent_qs}
\alias{find_adjacent_qs}
\title{Find adjacent quadrats}
\usage{
find_adjacent_qs(q, corner = TRUE)
}
\arguments{
\item{q}{character having length of 1 representing quadrat code to find.}

\item{corner}{if TRUE, includes quadrats touching at the corners of the quadrat.}
}
\value{
character vector representing codes of adjacent quadrats.
}
\description{
Find adjacent quadrats
}
