% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_error.r
\name{find_data_inconsistency}
\alias{find_data_inconsistency}
\title{Find data inconsistency for given ID}
\usage{
find_data_inconsistency(
  data,
  id_columns,
  data_columns = c("Q", "ind_id", "species", "x", "y"),
  deprecated_column = "修正済み"
)
}
\arguments{
\item{data}{a data.frame to be checked.}

\item{id_columns}{a character vector representing the name(s) of column(s) used as ID(s).}

\item{data_columns}{a character vector specifying the names of columns to check
consistency.}

\item{deprecated_column}{a character specifying the name of column showing deprecated records.}
}
\description{
Find records having multiple values for a given ID, such as different
quadrat codes, individual ids, species, etc.
}
