% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_data_error.r
\name{find_duplicated_stem_id}
\alias{find_duplicated_stem_id}
\title{Find duplicated stem ID in the stem sheet}
\usage{
find_duplicated_stem_id(
  stem_data,
  stem_id_column = "stem_id",
  deprecated_column = "修正済み",
  stem_level_info_columns = c("Q", "ind_id", "species")
)
}
\arguments{
\item{stem_data}{a data.frame having stem sheet.}

\item{stem_id_column}{a character specifying the column name of stem ID.}

\item{deprecated_column}{a character vector specifying the column name(s) of stem level
information by which inconsistency of stem level information
will be checked.}

\item{stem_level_info_columns}{column names of stem level information such as individual id or quadrat
code.}
}
\value{
A list having following fields:
*with_error*:
    a data.frame for the stem data with duplicated stem ID with some
    inconsistency of stem-level information.
*without_error*:
    a data.frame for the stem data with duplicated stem ID without
    inconsistent stem-level information.
 TODO: この関数、幹ID確認と幹レベル情報の確認に分離した方がよい。
}
\description{
Find duplicated stem ID from the stem sheet.
Also this function checks the duplicated ID(s) having inconsistent data
of stem level information.
}
