% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data_error.r
\name{find_excess_dead_records}
\alias{find_excess_dead_records}
\title{Find excess dead records}
\usage{
find_excess_dead_records(
  census_data,
  stem_id_column = "stem_id",
  deprecated_column = "ä¿®æ­£æ¸ˆã¿",
  ld_column = "ld",
  q_column = "Q",
  gbh_column = "gbh"
)
}
\arguments{
\item{census_data}{a data.frame having census data.}

\item{stem_id_column}{a character specifying the column name of stem ID.}

\item{deprecated_column}{a character specifying the name of the column having flags
of deprecated records.}

\item{ld_column}{a character specifying the name of the column alive/dead/unknown
status.}

\item{q_column}{a character specifying the name of the quadrat data.}

\item{gbh_column}{a character specifying the name of the gbh data.}
}
\value{
list of data.frames for stems with excess dead records.
}
\description{
Find stems having dead/unknown records >3 (1.2ha) or 2> (6ha).
}
