% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stem_data_error.r
\name{find_missing_ids}
\alias{find_missing_ids}
\title{Find missing stem and individual IDs.}
\usage{
find_missing_ids(
  stem_data,
  stem_id_column = "stem_id",
  individual_id_column = "ind_id"
)
}
\arguments{
\item{stem_data}{a data.frame having stem sheet.}

\item{stem_id_column}{a character specifying the column name of stem ID.}

\item{individual_id_column}{a character specifying the column name of individual ID.}
}
\description{
Find missing stem and individual IDs.
}
