% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data_error.r
\name{find_multiple_measurements}
\alias{find_multiple_measurements}
\title{Find multiple measurement}
\usage{
find_multiple_measurements(
  census_data,
  year_column = "year",
  deprecated_column = "修正済み",
  id_columns = c("stem_id", "tag_no", "Aタグ")
)
}
\arguments{
\item{census_data}{a data.frame having census data.}

\item{year_column}{a character specifying the column name of year.}

\item{deprecated_column}{a character specifying the column name of deprecated records.}

\item{id_columns}{character vector specifying the name(s) of column(s) containing ID(s).}
}
\value{
A list having records of multiple measurements detected for each ID
    specified as id_columns.
}
\description{
Find multiple measurement for each ID in each year.
}
