% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data_error.r
\name{find_resurrection}
\alias{find_resurrection}
\title{Find resurrected stems}
\usage{
find_resurrection(
  census_data,
  stem_id_column = "stem_id",
  deprecated_column = "修正済み",
  ld_column = "ld",
  year_column = "year",
  census_column = "s_date"
)
}
\arguments{
\item{census_data}{a data.frame having census data.}

\item{stem_id_column}{a character specifying the column name of stem ID.}

\item{deprecated_column}{a character specifying the column name of deprecated records.}

\item{ld_column}{a character specifying the column name of alive/dead/unknown status.}

\item{year_column}{a character specifying the column name of year.}

\item{census_column}{a character specifying the column name of census date.}
}
\value{
a list of data.frames for resurrected stems.
}
\description{
Find resurrected stems, i.e., L -> U -> L or L -> D -> L.
}
