% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data_error.r
\name{find_tags_on_different_stems}
\alias{find_tags_on_different_stems}
\title{Find same tag on different stems}
\usage{
find_tags_on_different_stems(
  census_data,
  stem_id_column = "stem_id",
  tag_names = c("tag_no", "Aタグ"),
  deprecated_column = "修正済み"
)
}
\arguments{
\item{census_data}{a data.frame having census data.}

\item{stem_id_column}{a character specifying the column name of stem ID.}

\item{tag_names}{a character vector specifying the column names of tags.}

\item{deprecated_column}{a character specifying the column name of deprecated records.}
}
\description{
Find same tag on different stems
}
