% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_data_error.r
\name{find_unknown_with_measurements}
\alias{find_unknown_with_measurements}
\title{Find records with unknown and measured values}
\usage{
find_unknown_with_measurements(
  census_data,
  stem_id_column = "stem_id",
  ld_column = "ld",
  deprecated_column = "修正済み",
  year_column = "year",
  unknown_code = "U",
  measurements = c("gbh", "cls")
)
}
\arguments{
\item{census_data}{census data.}

\item{stem_id_column}{a character specifying the column name of stem ID.}

\item{ld_column}{a character specifying the column name of alive/dead/unknown status.}

\item{deprecated_column}{a character specifying the column name of deprecated records.}

\item{year_column}{a character specifying the column name of year.}

\item{unknown_code}{a character specifying unknown code.}

\item{measurements}{a character vector specifying the column names having measurements.}
}
\value{
a data.frame having records of stems having both unknown and
    measurements in a year.
    Returns NULL if no record was detected.
}
\description{
Find records of a stem in a year with unknown and measured value(s).
}
