% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{invalid_xy}
\alias{invalid_xy}
\title{Validate coordinates}
\usage{
invalid_xy(sq2, x, y)
}
\arguments{
\item{sq2}{last digit of quadrat code.}

\item{x}{x-coordinate.}

\item{y}{y coordinate.}
}
\value{
TRUE if the coordinate is not valid.
}
\description{
Validate coordinates
}
