% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is.adjacent}
\alias{is.adjacent}
\title{Check if the plot pairs are adjacent}
\usage{
is.adjacent(q1, q2)
}
\arguments{
\item{q1}{quadrat codes, should have equal length with q2.}

\item{q2}{quadrat codes, should have equal length with q1.}
}
\value{
Returns TRUE if the two plots are adjacent and FALSE if are not.
}
\description{
Check if the plot pairs are adjacent
}
\examples{
q1 <- c("A1a2", "A1a2", "A1a1")
q2 <- c("A1a4", "B1a2", "A1a3")
is.adjacent(q1, q2)

}
