% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is.core}
\alias{is.core}
\title{Is core plot?}
\usage{
is.core(Q, invalid = NA)
}
\arguments{
\item{Q}{vector of quadrat code(s), e.g., "A1b2" or "B02c3".}

\item{invalid}{a value returned for invalid quadrat codes including empty characters and
NAs.}
}
\value{
logical indicating whether specified quadrats are in 1.2ha region.
For NAs and empty characters, return NA.
}
\description{
Determine whether the quadrat is in the 1.2ha  plot.
}
