% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is_in}
\alias{is_in}
\title{Check if quadrats are in a specified region}
\usage{
is_in(Q, xmin = NA, xmax = NA, ymin = NA, ymax = NA, q_from = NA, q_to = NA)
}
\arguments{
\item{Q}{quadrate codes.}

\item{xmin}{minimum x-coordinate of the region.}

\item{xmax}{maximum x-coordinate of the region.}

\item{ymin}{minimum y-coordinate of the region.}

\item{ymax}{maximum y-coordinate of the region.}

\item{q.from}{a quadrat code specifying the region begins.}

\item{q.to}{a quadrat code specifying the region ends.}
}
\value{
a logical vector indicating if the quadrats are in the region.
}
\description{
Check if quadrats are in a specified region.
The region can be specified by using minimum and maximum coordinates or two
quadrats indicating starting from and to.
}
\examples{
is.in("A1a1", 0, 100, 0, 200)
is.in("A1a1", q.from = "B1a1", q.to = "B3c2")

}
