% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generic_error.r
\name{is_valid_1cm_tag}
\alias{is_valid_1cm_tag}
\title{Validate format of 1cm tag}
\usage{
is_valid_1cm_tag(tag_no)
}
\arguments{
\item{tag_no}{a character vector containing 1cm tag numbers.}
}
\value{
for valid 1cm tag numbers, return TRUE.
}
\description{
Validate format of 1cm tag
}
