% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{is_valid_q}
\alias{is_valid_q}
\title{Validate quadrat codes}
\usage{
is_valid_q(Q)
}
\arguments{
\item{Q}{quadrat code(s).}
}
\value{
returns TRUE for quadrat codes in valid format and FALSE for invalid
    format.
}
\description{
Validate quadrat codes
}
