% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.r
\name{point_q}
\alias{point_q}
\title{Draw quadrats as points with/without jitter.}
\usage{
point_q(Q, jitter = FALSE, ...)
}
\arguments{
\item{Q}{quadrat codes.}

\item{jitter}{if TRUE, add jitter for coordinates of overlapping points.}

\item{...}{graphic parameters passed to \code{\link[graphics]{points}}.}
}
\description{
Draw quadrats as points with/without jitter.
}
