% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.r
\name{point_rc}
\alias{point_rc}
\title{Draw row and column as points with/without jitter.}
\usage{
point_rc(R, C, jitter = FALSE, ...)
}
\arguments{
\item{R}{row.}

\item{C}{column.}

\item{jitter}{if TRUE, add jitter for coordinates of overlapping points.}

\item{...}{graphic parameters passed to \code{\link[graphics]{points}}.}
}
\description{
Draw row and column as points with/without jitter.
}
