% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{q.to.elements}
\alias{q.to.elements}
\title{Split quadrat codes into elements}
\usage{
q.to.elements(Q)
}
\arguments{
\item{Q}{vector of quadrat code(s).}
}
\value{
a list of quadrat code elements.
For invalit quadrat codes, NAs are returned.
}
\description{
Split quadrat codes into elements
}
