% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{q.to.point}
\alias{q.to.point}
\title{Convert quadrat codes to coordinates of points}
\usage{
q.to.point(
  Q,
  Q1 = NA,
  Q2 = NA,
  SQ1 = NA,
  SQ2 = NA,
  pos = c("center", "topleft", "topright", "bottomleft", "bottomright")
)
}
\arguments{
\item{Q}{vector of quadrat code(s). When Q is specified, Q1, Q2, SQ1 and Q2 are
ignored.}

\item{Q1}{vector of capital alphabet part of quadrat code.}

\item{Q2}{vector of first numerical part of quadrat code.}

\item{SQ1}{vector of lower alphabet part of quadrat code.}

\item{SQ2}{vector of second numerical part of quadrat code.}

\item{pos}{position of the quadrat, can be one of "center", "topleft", "topright",
"bottomleft", "bottomright".}
}
\value{
list of x and y coordinates of the point(s).
}
\description{
Convert quadrat codes to coordinates of points
}
