% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{q.to.rect}
\alias{q.to.rect}
\title{Convert quadrat code to coordinates of rectangular for the quadrat}
\usage{
q.to.rect(Q, Q1 = NA, Q2 = NA, SQ1 = NA, SQ2 = NA)
}
\arguments{
\item{Q}{vector of quadrat code(s).}

\item{Q1}{vector of capital alphabet part of quadrat code.}

\item{Q2}{vector of first numerical part of quadrat code.}

\item{SQ1}{vector of lower alphabet part of quadrat code.}

\item{SQ2}{vector of second numerical part of quadrat code.}
}
\value{
list of coordinates (x1, x2, y1, y2) for the rectangular.
}
\description{
Convert quadrat code to coordinates of rectangular for the quadrat
}
