% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{q.to.x}
\alias{q.to.x}
\title{Convert quadrat codes to x coordinates}
\usage{
q.to.x(Q, Q1 = NA, Q2 = NA, SQ1 = NA, SQ2 = NA)
}
\arguments{
\item{Q}{vector of quadrat code(s).}

\item{Q1}{vector of capital alphabet part of quadrat code.}

\item{Q2}{vector of first numerical part of quadrat code.}

\item{SQ1}{vector of lower alphabet part of quadrat code.}

\item{SQ2}{vector of second numerical part of quadrat code.}
}
\value{
a vector of x-coordinates.
}
\description{
Convert quadrat codes to x coordinates of top-right of the quadrats.
}
\seealso{
\code{\link{q.to.y}}
}
