% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{q_to_elements}
\alias{q_to_elements}
\title{Split quadrat codes into elements}
\usage{
q_to_elements(Q)
}
\arguments{
\item{Q}{vector of quadrat code(s).}
}
\value{
a list of quadrat code elements.
    For invalid quadrat codes, NAs are returned.
}
\description{
Split quadrat codes into elements
}
