% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{q_to_x}
\alias{q_to_x}
\title{Convert quadrat codes to x coordinates}
\usage{
q_to_x(Q, Q1, Q2 = NA, SQ1 = NA, SQ2 = NA)
}
\arguments{
\item{Q}{vector of quadrat code(s).}

\item{Q1}{vector of capital alphabet part of quadrat code.}

\item{Q2}{vector of first numerical part of quadrat code.
Can be specified for intuitiveness, but not used for any calculation.}

\item{SQ1}{vector of lower alphabet part of quadrat code.}

\item{SQ2}{vector of second numerical part of quadrat code.}
}
\value{
a vector of x-coordinates.
}
\description{
Convert quadrat codes to x coordinates of bottom-left of the quadrats.
When SQ2 is not specified, returns coordinates of 10m grid.
When SQ1 is not specified, returns coordinates of 20m grid.
}
\seealso{
\code{\link{q_to_y}}
}
