% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{standardize.q}
\alias{standardize.q}
\title{Standardize quadrat codes}
\usage{
standardize.q(Q)
}
\arguments{
\item{Q}{vector of quadrat code.}
}
\value{
a character vector of quadrate code(s).
}
\description{
Convert quadrat code(s) in the A1a1 format to the A01a1 format.
}
