% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{to_plot_x}
\alias{to_plot_x}
\title{Convert x coordinate within a quadrat to global quadrat within the plot}
\usage{
to_plot_x(x, Q, Q1, Q2 = NA, SQ1 = NA, SQ2 = NA)
}
\arguments{
\item{x}{x coordinate within a quadrat.}

\item{Q}{vector of quadrat code(s).}

\item{Q1}{vector of capital alphabet part of quadrat code.}

\item{Q2}{vector of first numerical part of quadrat code.}

\item{SQ1}{vector of lower alphabet part of quadrat code.}

\item{SQ2}{vector of second numerical part of quadrat code.}
}
\value{
vector of x coordinates. When x is NA, returns center of the quadrat.
}
\description{
Convert x coordinate within a quadrat to global quadrat within the plot
}
