% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotmap.r
\name{create.ogawa.plot}
\alias{create.ogawa.plot}
\title{Create new ogawa forest plot}
\usage{
create.ogawa.plot(xmin = 0, xmax = 300, ymin = 0, ymax = 200)
}
\arguments{
\item{xmin}{a minimum value of x coordinate.}

\item{xmax}{a maximum value of x coordinate.}

\item{ymin}{a minimum value of y coordinate.}

\item{ymax}{a maximum value of y coordinate.}
}
\value{
an \code{ogara.plot} object which can be passed to other functions.
}
\description{
Create new plot for ogawa forest plot map.
}
\examples{
x <- create.ogawa.plot()
add.grid(x)

}
